refactor: shuffle types
