+++ packages/sdk/wallets/wallet-walletconnect/package.json
@@ -18,7 +18,7 @@
     "clean": "tsc -b . --clean",
     "docs": "typedoc && ts-node ../../utils/scripts/linkdocs.ts wallet-walletconnect",
     "test": "jest --runInBand",
-    "test-staging": "WALLET_CONNECT_BRIDGE=wss://relay.walletconnect.org yarn test",
+    "test-staging": "WALLET_CONNECT_BRIDGE=wss://walletconnect.celo-networks-dev.org yarn test",
     "lint": "tslint -c tslint.json --project .",
     "prepublishOnly": "yarn build"
   },
@@ -28,6 +28,7 @@
     "@celo/wallet-base": "1.1.1-dev",
     "@celo/wallet-remote": "1.1.1-dev",
     "@walletconnect/client": "2.0.0-alpha.35",
+    "@walletconnect/types": "2.0.0-alpha.35",
     "debug": "^4.1.1",
     "ethereumjs-util": "^7.0.8"
   },
